FileTransfer_module/requirements.txt
boto3==1.26.0 
google-cloud-storage==2.11.0  
python-dotenv==1.0.0      
